Sswissbit incrememntal export from tse

+++ scu-de/src/fiskaltrust.Middleware.SCU.DE.Swissbit/SwissbitSCU.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
@@ -11,6 +12,7 @@ using System.Threading.Tasks;
 using fiskaltrust.ifPOS.v1.de;
 using fiskaltrust.Middleware.SCU.DE.Helpers.TLVLogParser.Logs;
 using fiskaltrust.Middleware.SCU.DE.Helpers.TLVLogParser.Logs.Models;
+using fiskaltrust.Middleware.SCU.DE.Helpers.TLVLogParser.Tar;
 using fiskaltrust.Middleware.SCU.DE.Swissbit.Exceptions;
 using fiskaltrust.Middleware.SCU.DE.Swissbit.Helpers;
 using fiskaltrust.Middleware.SCU.DE.Swissbit.Interop;
@@ -42,6 +44,7 @@ namespace fiskaltrust.Middleware.SCU.DE.Swissbit
         private uint _hwSelftestIntervalSeconds = 0;
         private ISwissbitProxy _proxy = null;
         private DateTime _nextSyncTime;
+        private ulong _lastTransactionNr = 0;
 
         // Never change these values, as all existing installations are depending on them
         private readonly byte[] _adminPuk = Encoding.ASCII.GetBytes("123456");
@@ -598,6 +601,7 @@ namespace fiskaltrust.Middleware.SCU.DE.Swissbit
                                 SignatureCounter = tseResponse.SignatureCounter
                             }
                         };
+                        _lastTransactionNr = tseResponse.TransactionNumber;
 
                         return response;
                     }
@@ -738,7 +742,20 @@ namespace fiskaltrust.Middleware.SCU.DE.Swissbit
 
                 await UpdateTimeAsync(GetProxy());
                 SetExportState(exportId, ExportState.Running);
-                CacheExportAsync(exportId, request.ClientId, request.Erase).FireAndForget();
+
+                if (_configuration.ChunkExportTransactionCount > 0)
+                {
+                    if (_lastTransactionNr == 0)
+                    {
+                        _logger.LogWarning("Before executing a partial export a daily closing has to be made.");
+                        throw new Exception("Missing Daily Closing.");
+                    };
+                    CacheExportIncrementalAsync(exportId, 0, _configuration.ChunkExportTransactionCount, (long) _lastTransactionNr).FireAndForget();
+                }
+                else
+                {
+                    CacheExportAsync(exportId, request.ClientId, request.Erase).FireAndForget();
+                }
 
                 return new StartExportSessionResponse()
                 {
@@ -782,6 +799,73 @@ namespace fiskaltrust.Middleware.SCU.DE.Swissbit
                     SetExportState(exportId, ExportState.Failed, ex);
                 }
             });
+        }
+
+        private async Task CacheExportIncrementalAsync(Guid exportId, long startTransactionNr, long maxNumberOfRecords, long currentNumberOfTransactions)
+        {
+            var endTransactionNr = LastNumberOfRecords(startTransactionNr, maxNumberOfRecords, currentNumberOfTransactions);
+            long newStartTransactionNr = -1;
+
+            await _lockingHelper.PerformWithLock(_hwLock, async () =>
+            {
+                try
+                {
+                    using (var stream = new MemoryStream())
+                    {
+                        await GetProxy().ExportTarFilteredTransactionAsync(stream, (ulong) startTransactionNr, (ulong) endTransactionNr, null);
+                        stream.Position = 0;
+                        TarFileHelper.AppendTarStreamToTarFile(exportId.ToString(), stream);
+                        _logger.LogInformation($"Export total {currentNumberOfTransactions}. Partial Export from TransactionNr: {startTransactionNr} to TransactionNr: {endTransactionNr}");
+                    }
+                    newStartTransactionNr = GetLastExportedTransaction(exportId.ToString()) + 1;
+                }
+                catch (Exception ex)
+                {
+                    if (!ex.Message.Equals("Filtered Export: no matching entries, export would be empty. "))
+                    {
+                        _logger.LogError(ex, "Failed to execute {Operation} - TempFileName: {TempFileName}", nameof(CacheExportIncrementalAsync), exportId.ToString());
+                        SetExportState(exportId, ExportState.Failed, ex);
+                        throw;
+                    }
+                    _logger.LogInformation($"No Data: Export total {currentNumberOfTransactions}. Partial Export from TransactionNr: {startTransactionNr} to TransactionNr: {endTransactionNr}");
+
+                }
+            });
+
+            if (newStartTransactionNr == -1)
+            {
+                newStartTransactionNr = startTransactionNr + maxNumberOfRecords;
+            }
+            if (newStartTransactionNr <= currentNumberOfTransactions)
+            {
+                await CacheExportIncrementalAsync(exportId, newStartTransactionNr, maxNumberOfRecords, currentNumberOfTransactions).ConfigureAwait(false);
+            }
+            else
+            {
+                TarFileHelper.FinalizeTarFile(exportId.ToString());
+                _logger.LogDebug("Finalized merged TAR file {fileName}.", exportId.ToString());
+                SetExportState(exportId, ExportState.Succeeded);
+            }
+        }
+
+        private long GetLastExportedTransaction(string targetFile)
+        {
+            var lastlog = TarFileHelper.GetLastLogEntryFromTarFile(targetFile);
+
+            //Unixt_1687767355_Sig-105945_Log-Tra_No-50549_Finish_Client-6qmZrXEtrkuBhSOA7Oi39g.log
+            var iSigStart = lastlog.IndexOf("Tra_No-")+7;
+            var iSigEnd = lastlog.IndexOf('_', iSigStart);
+            var lastSigCount = lastlog.Substring(iSigStart, iSigEnd - iSigStart);
+            return long.Parse(lastSigCount);
+        }
+
+        private long LastNumberOfRecords(long previousSignatureCounter, long maxNumberOfRecords, long currentNumberOfSignatures)
+        {
+            if ((previousSignatureCounter + maxNumberOfRecords) > currentNumberOfSignatures)
+            {
+                return currentNumberOfSignatures;
+            }
+            return previousSignatureCounter + maxNumberOfRecords - 1;
         }
 
         private void SetEraseEnabledForExportState(Guid tokenId, ExportState exportState, Exception error = null)
@@ -949,7 +1033,7 @@ namespace fiskaltrust.Middleware.SCU.DE.Swissbit
                             }
                             try
                             {
-                                if (File.Exists(tempFileName))
+                                if (File.Exists(tempFileName) && !_configuration.StoreTemporaryFile)
                                 {
                                     File.Delete(tempFileName);
                                 }
